refactor utilities.py > utility.py
